// src/components/button/enum.ts
export enum ButtonState {
	PRIMARY = "bg-primary-400 rounded flex justify-center items-center text-white font-semibold text-sm-regular my-1",
	SECONDARY = "bg-white rounded border flex justify-center items-center text-primary-400 font-semibold text-sm-regular my-1",
	TERTIARY = "bg-neutral-100 rounded flex justify-center items-center text-neutral-300 font-semibold text-sm-regular my-1",
	DISABLED = "bg-gray-200 rounded flex justify-center items-center text-gray-300 font-semibold text-sm-regular my-1",
	OUTLINE = "rounded flex justify-center items-center border text-primary-300 border-primary-300 rounded text-sm-regular",
	SUCCESS = "bg-active rounded flex justify-center items-center text-white font-medium text-sm-regular my-1",
	ORANGE = "bg-orange rounded flex justify-center items-center text-white font-medium text-sm-regular my-1",
	NEUTRAL = "bg-orange rounded flex justify-center items-center text-white font-medium text-sm-regular my-1",
	PRIMARYALT = "rounded flex justify-center items-center text-primary-400 bg-white font-semibold text-sm-regular my-1",
	OUTLINEALT = "rounded flex justify-center items-center border text-white border-white  text-sm-regular my-1",
	SKIP = "rounded flex justify-center  items-center  text-primary-400 text-sm-regular my-1 bg-gray-fade ",
	GRAYALT = "rounded flex justify-center items-center text-neutral-400 text-sm-regular my-1 border border-gray-100 bg-white",
	NEUTRALP2P = "font-medium border-x border-neutral-60 text-neutral-400 rounded my-1",
}

export enum ButtonSize {
	lg = "py-3.5 px-6",
	sm = "py-2.5 px-3",
	xs = "py-1 px-1.5",
}